move magic energy to Shared, add manacost to spells
