+++ Content.Server/_CP14/MagicEnergy/CP14MagicEnergySystem.cs
@@ -1,11 +1,8 @@
 using Content.Server._CP14.MagicEnergy.Components;
 using Content.Shared._CP14.MagicEnergy;
 using Content.Shared._CP14.MagicEnergy.Components;
-using Content.Shared.Alert;
 using Content.Shared.Examine;
-using Content.Shared.FixedPoint;
 using Content.Shared.Inventory;
-using Content.Shared.Rounding;
 using Robust.Server.GameObjects;
 using Robust.Shared.Timing;
 
@@ -13,17 +10,12 @@ namespace Content.Server._CP14.MagicEnergy;
 
 public sealed partial class CP14MagicEnergySystem : SharedCP14MagicEnergySystem
 {
-    [Dependency] private readonly AlertsSystem _alerts = default!;
     [Dependency] private readonly IGameTiming _gameTiming = default!;
     [Dependency] private readonly PointLightSystem _light = default!;
     [Dependency] private readonly CP14MagicEnergyCrystalSlotSystem _magicSlot = default!;
 
     public override void Initialize()
     {
-
-        SubscribeLocalEvent<CP14MagicEnergyContainerComponent, ComponentStartup>(OnComponentStartup);
-        SubscribeLocalEvent<CP14MagicEnergyContainerComponent, ComponentShutdown>(OnComponentShutdown);
-
         SubscribeLocalEvent<CP14MagicEnergyDrawComponent, MapInitEvent>(OnMapInit);
 
         SubscribeLocalEvent<CP14MagicEnergyPointLightControllerComponent, CP14MagicEnergyLevelChangeEvent>(OnEnergyChange);
@@ -31,19 +23,6 @@ public sealed partial class CP14MagicEnergySystem : SharedCP14MagicEnergySystem
         SubscribeLocalEvent<CP14MagicEnergyExaminableComponent, ExaminedEvent>(OnExamined);
         SubscribeLocalEvent<CP14MagicEnergyScannerComponent, CP14MagicEnergyScanEvent>(OnMagicScanAttempt);
         SubscribeLocalEvent<CP14MagicEnergyScannerComponent, InventoryRelayedEvent<CP14MagicEnergyScanEvent>>((e, c, ev) => OnMagicScanAttempt(e, c, ev.Args));
-    }
-
-    private void OnComponentStartup(Entity<CP14MagicEnergyContainerComponent> ent, ref ComponentStartup args)
-    {
-        UpdateMagicAlert(ent);
-    }
-
-    private void OnComponentShutdown(Entity<CP14MagicEnergyContainerComponent> ent, ref ComponentShutdown args)
-    {
-        if (ent.Comp.MagicAlert == null)
-            return;
-
-        _alerts.ClearAlert(ent, ent.Comp.MagicAlert.Value);
     }
 
     private void OnEnergyChange(Entity<CP14MagicEnergyPointLightControllerComponent> ent, ref CP14MagicEnergyLevelChangeEvent args)
@@ -79,8 +58,6 @@ public sealed partial class CP14MagicEnergySystem : SharedCP14MagicEnergySystem
         args.PushMarkup(GetEnergyExaminedText(ent, magicContainer));
     }
 
-
-
     public override void Update(float frameTime)
     {
         base.Update(frameTime);
@@ -112,73 +89,5 @@ public sealed partial class CP14MagicEnergySystem : SharedCP14MagicEnergySystem
 
             ChangeEnergy(energyEnt.Value, energyComp, draw.Energy, draw.Safe);
         }
-    }
-
-    public bool TryConsumeEnergy(EntityUid uid, FixedPoint2 energy, CP14MagicEnergyContainerComponent? component = null, bool safe = false)
-    {
-        if (!Resolve(uid, ref component))
-            return false;
-
-        if (energy <= 0)
-            return true;
-
-        // Attempting to absorb more energy than is contained in the carrier will still waste all the energy
-        if (component.Energy < energy)
-        {
-            ChangeEnergy(uid, component, -component.Energy);
-            return false;
-        }
-
-        ChangeEnergy(uid, component, -energy, safe);
-        return true;
-    }
-
-    public void ChangeEnergy(EntityUid uid, CP14MagicEnergyContainerComponent component, FixedPoint2 energy, bool safe = false)
-    {
-        if (!safe)
-        {
-            //Overload
-            if (component.Energy + energy > component.MaxEnergy)
-            {
-                RaiseLocalEvent(uid, new CP14MagicEnergyOverloadEvent()
-                {
-                    OverloadEnergy = (component.Energy + energy) - component.MaxEnergy,
-                });
-            }
-
-            //Burn out
-            if (component.Energy + energy < 0)
-            {
-                RaiseLocalEvent(uid, new CP14MagicEnergyBurnOutEvent()
-                {
-                    BurnOutEnergy = -energy - component.Energy
-                });
-            }
-        }
-
-        var oldEnergy = component.Energy;
-        var newEnergy = Math.Clamp((float)component.Energy + (float)energy, 0, (float)component.MaxEnergy);
-        component.Energy = newEnergy;
-
-        if (oldEnergy != newEnergy)
-        {
-            RaiseLocalEvent(uid, new CP14MagicEnergyLevelChangeEvent()
-            {
-                OldValue = component.Energy,
-                NewValue = newEnergy,
-                MaxValue = component.MaxEnergy,
-            });
-        }
-
-        UpdateMagicAlert((uid, component));
-    }
-
-    private void UpdateMagicAlert(Entity<CP14MagicEnergyContainerComponent> ent)
-    {
-        if (ent.Comp.MagicAlert == null)
-            return;
-
-        var level = ContentHelpers.RoundToLevels(MathF.Max(0f, (float) ent.Comp.Energy), (float) ent.Comp.MaxEnergy, 10);
-        _alerts.ShowAlert(ent, ent.Comp.MagicAlert.Value, (short)level);
     }
 }
