+++ Resources/Locale/en-US/_CP14/magicEnergy/magic-spells.ftl
@@ -1,0 +1,1 @@
+cp14-magic-spell-not-enough-mana = Not enough mana!
