+++ Resources/Prototypes/_CP14/Entities/Actions/test_spell.yml
@@ -32,7 +32,7 @@
   description: Fires an explosive fireball towards the clicked location.
   components:
   - type: CP14MagicEffect
-    manaCost: 10
+    manaCost: 70
   - type: WorldTargetAction
     useDelay: 1
     itemIconStyle: BigAction
@@ -74,6 +74,8 @@
   name: Gravity well
   description: Toggles your suit's phase cloak. Beware that if you are hit, all abilities are disabled for 5 seconds, including your cloak!
   components:
+  - type: CP14MagicEffect
+    manaCost: 20
   - type: WorldTargetAction
     useDelay: 1
     itemIconStyle: BigAction
